Create fibonacci.js
Three versions of fibonacci sequence computation. One is purely iterative with a sotrage of all the numbers within an array. The second is recursive, but has to calculate all of the sub computations of every recursive branch every time. The third is a recursive function that tries to minimize memory overhead.